Edits to Survey Rules
- Remove toggleState method
- Renamed SurveyRules to SurveyQuestionRules
- Use SvgIcon for close icon
